fix compilation on gcc 4.5
GCC 4.5 does more template instantiations than earlier, for standard
conformance reasons. The net result was that some invalid MPL
specializations *where* instantiated and rejected and are now
instantiated *and* generate errors.

Description of the problem and fix:

  http://gcc.gnu.org/bugzilla/show_bug.cgi?id=43608

This commit implements the "do explicit specialization" fix
proposed in the thread. For that, I needed to move the MPL stuff
in its own namespace (as explicit template specialization cannot
be done at class level).